Refactor to move test helpers to create mock apps.
Move them into the `utils` module so that they can be used by other
integration tests.
