+++ linera-execution/tests/test_execution.rs
@@ -5,27 +5,23 @@
 
 mod utils;
 
-use self::utils::{create_dummy_user_application_description, ExpectedCall, MockApplication};
+use self::utils::{
+    create_dummy_user_application_registrations, register_mock_applications, ExpectedCall,
+};
 use linera_base::{
     crypto::PublicKey,
     data_types::BlockHeight,
     identifiers::{ChainDescription, ChainId, Destination, Owner},
 };
 use linera_execution::{
-    system::SystemMessage, ApplicationCallOutcome, ApplicationRegistryView, BaseRuntime,
-    ContractRuntime, ExecutionError, ExecutionOutcome, ExecutionRuntimeConfig,
-    ExecutionRuntimeContext, ExecutionStateView, MessageKind, Operation, OperationContext, Query,
-    QueryContext, RawExecutionOutcome, RawOutgoingMessage, ResourceController, Response,
-    SessionCallOutcome, SystemExecutionState, TestExecutionRuntimeContext,
-    UserApplicationDescription, UserApplicationId,
+    system::SystemMessage, ApplicationCallOutcome, BaseRuntime, ContractRuntime, ExecutionError,
+    ExecutionOutcome, ExecutionRuntimeConfig, ExecutionStateView, MessageKind, Operation,
+    OperationContext, Query, QueryContext, RawExecutionOutcome, RawOutgoingMessage,
+    ResourceController, Response, SessionCallOutcome, SystemExecutionState,
+    TestExecutionRuntimeContext,
 };
-use linera_views::{
-    batch::Batch,
-    common::Context,
-    memory::MemoryContext,
-    views::{View, ViewError},
-};
-use std::{sync::Arc, vec};
+use linera_views::{batch::Batch, memory::MemoryContext};
+use std::vec;
 
 #[tokio::test]
 async fn test_missing_bytecode_for_user_application() -> anyhow::Result<()> {
@@ -911,58 +907,4 @@ async fn test_multiple_messages_from_different_applications() -> anyhow::Result<
     );
 
     Ok(())
-}
-
-/// Creates `count` [`MockApplication`]s and registers them in the provided [`ExecutionStateView`].
-///
-/// Returns an iterator over pairs of [`UserApplicationId`]s and their respective
-/// [`MockApplication`]s.
-pub async fn register_mock_applications<C>(
-    state: &mut ExecutionStateView<C>,
-    count: u64,
-) -> anyhow::Result<vec::IntoIter<(UserApplicationId, MockApplication)>>
-where
-    C: Context<Extra = TestExecutionRuntimeContext> + Clone + Send + Sync + 'static,
-    ViewError: From<C::Error>,
-{
-    let mock_applications: Vec<_> =
-        create_dummy_user_application_registrations(&mut state.system.registry, count)
-            .await?
-            .into_iter()
-            .map(|(id, _description)| (id, MockApplication::default()))
-            .collect();
-    let extra = state.context().extra();
-
-    for (id, mock_application) in &mock_applications {
-        extra
-            .user_contracts()
-            .insert(*id, Arc::new(mock_application.clone()));
-        extra
-            .user_services()
-            .insert(*id, Arc::new(mock_application.clone()));
-    }
-
-    Ok(mock_applications.into_iter())
-}
-
-pub async fn create_dummy_user_application_registrations<C>(
-    registry: &mut ApplicationRegistryView<C>,
-    count: u64,
-) -> anyhow::Result<Vec<(UserApplicationId, UserApplicationDescription)>>
-where
-    C: Context + Clone + Send + Sync + 'static,
-    ViewError: From<C::Error>,
-{
-    let mut ids = Vec::with_capacity(count as usize);
-
-    for index in 0..count {
-        let description = create_dummy_user_application_description(index);
-        let id = registry.register_application(description.clone()).await?;
-
-        assert_eq!(registry.describe_application(id).await?, description);
-
-        ids.push((id, description));
-    }
-
-    Ok(ids)
 }
